# Reduce sort bucket exchange overhead