@mhmerrill 2022-07-28
> Drafting only because I'm not around if there's any fallout from this. I feel pretty confident about the change if you want to try it out, but if you want to wait I'll be back on the 8th.

i will also be out until Aug 8th.